Working on implementing JAMA. Why in the hell can't they have nice toStrings?

+++ MatchAnalysis/src/Inverse.java
@@ -8,13 +8,13 @@ public class Inverse
         Scanner input = new Scanner(System.in);
         System.out.println("Enter the dimension of square matrix: ");
         int n = input.nextInt();
-        double a[][]= new double[n][n];
+        int a[][]= new int[n][n];
         System.out.println("Enter the elements of matrix: ");
         for(int i=0; i<n; i++)
             for(int j=0; j<n; j++)
-                a[i][j] = input.nextDouble();
+                a[i][j] = input.nextInt();
  
-        double d[][] = invert(a);
+        int d[][] = invert(a);
  
         System.out.println("The inverse is: ");
         for (int i=0; i<n; ++i) 
@@ -28,11 +28,11 @@ public class Inverse
         input.close();
     }	
  
-    public static double[][] invert(double a[][]) 
+    public static int[][] invert(int a[][]) 
     {
         int n = a.length;
-        double x[][] = new double[n][n];
-        double b[][] = new double[n][n];
+        int x[][] = new int[n][n];
+        int b[][] = new int[n][n];
         int index[] = new int[n];
         for (int i=0; i<n; ++i) 
             b[i][i] = 1;
@@ -67,10 +67,10 @@ public class Inverse
 // Method to carry out the partial-pivoting Gaussian
 // elimination.  Here index[] stores pivoting order.
  
-    public static void gaussian(double a[][], int index[]) 
+    public static void gaussian(int a[][], int index[]) 
     {
         int n = index.length;
-        double c[] = new double[n];
+        int c[] = new int[n];
  
  // Initialize the index
         for (int i=0; i<n; ++i) 
@@ -79,10 +79,10 @@ public class Inverse
  // Find the rescaling factors, one from each row
         for (int i=0; i<n; ++i) 
         {
-            double c1 = 0;
+            int c1 = 0;
             for (int j=0; j<n; ++j) 
             {
-                double c0 = Math.abs(a[i][j]);
+                int c0 = Math.abs(a[i][j]);
                 if (c0 > c1) c1 = c0;
             }
             c[i] = c1;
@@ -92,10 +92,10 @@ public class Inverse
         int k = 0;
         for (int j=0; j<n-1; ++j) 
         {
-            double pi1 = 0;
+            int pi1 = 0;
             for (int i=j; i<n; ++i) 
             {
-                double pi0 = Math.abs(a[index[i]][j]);
+                int pi0 = Math.abs(a[index[i]][j]);
                 pi0 /= c[index[i]];
                 if (pi0 > pi1) 
                 {
@@ -110,7 +110,7 @@ public class Inverse
             index[k] = itmp;
             for (int i=j+1; i<n; ++i) 	
             {
-                double pj = a[index[i]][j]/a[index[j]][j];
+                int pj = a[index[i]][j]/a[index[j]][j];
  
  // Record pivoting ratios below the diagonal
                 a[index[i]][j] = pj;
